+++ packages/date-ui/.flowconfig
@@ -1,0 +1,12 @@
+[ignore]
+<PROJECT_ROOT>/node_modules/radium/.*
+
+[include]
+
+[libs]
+
+[options]
+esproposal.class_static_fields=enable
+esproposal.class_instance_fields=enable
+
+[lints]
