+++ packages/date-ui/.gitignore
@@ -1,0 +1,6 @@
+node_modules
+dist
+lib
+coverage
+docs/*
+!docs/README.md
