+++ packages/date-ui/.npmignore
@@ -1,0 +1,14 @@
+coverage/
+deprecated/
+docs/
+node_modules/
+src/
+tests/
+.flowconfig
+.gitignore
+.npmignore
+jsdoc.globals.js
+jsdoc.json
+index.server.js
+webpack.config.babel.js
+dist/
