+++ packages/date-ui/jsdoc.globals.js
@@ -1,0 +1,24 @@
+/**
+ * React component
+ * @typedef Component
+ * @global
+ */
+
+/**
+ * React element
+ * @typedef Element
+ * @global
+ */
+
+/**
+ * Standard onChange handler. Called when data in a component changes.
+ * @global
+ * @callback onChange
+ * @param {*} data - The changed data.
+ */
+
+/**
+ * Redux Action Object - An object with a type that is sent to Redux reducers.
+ * @global
+ * @typedef ReduxAction
+ */
