+++ packages/date-ui/src/util.js
@@ -1,0 +1,8 @@
+import momentLocalizer from 'react-widgets-moment';
+
+/**
+ * Initializes the date-time pickers
+ */
+export function dateInit() {
+	momentLocalizer();
+}
