+++ packages/date-ui/tests/__mocks__/fileMock.js
