+++ packages/date-ui/webpack.config.babel.js
